feat: add jump to room feature and admin dashboard with authentication (#11)
* feat: add jump to room feature and selection color

* feat: add admin dashboard with jwt authentication

* chore: move files to map folder

* chore: refactor error handling and holiday parsing in services

Added error handling for OSMExtractorService and improved parsing logic for holiday events in StarPlanService. Extended test cases and mock data to validate the changes. Updated configuration with additional building mappings and admin IDs.

* chore: add credits and found error popup

* chore: reorder navigation components

* chore: add github link

* chore: add tests and remove duplicate code

* chore: add tests

* Add initial tests for RoomClicker and Admin components

This commit introduces tests for the RoomClicker and Admin components using Vitest and React Testing Library. Mocking is set up for required dependencies, and basic rendering and interaction scenarios are validated. Some tests are marked as TODO for future enhancement.

* fix: fix lint issues

* chore: add RoomContext test

* chore: add AdminServiceTest.kt

* chore: add first campus map test

* chore: add AdminControllerTest.kt

* chore: add more tests

* chore: add feedback to admin dashboard